#86401 write tests for extensions merge
